fix: int cast before === User::AUTH_PUBLIC
`User::AUTH_PUBLIC` is an int, so the config value needs to be cast to
an int before checking with `===`-operator.

This fixes a problem where having `GRAPHER_ACCESS_CUSTOMER=0` in .env is
not giving public access. Probably fixes other cases too.

+++ app/Services/Grapher/Graph/Latency.php
@@ -241,7 +241,7 @@ class Latency extends Graph
             return true;
         }
 
-        if( !Auth::check() && is_numeric( config( 'grapher.access.latency' ) ) && config( 'grapher.access.latency' ) === User::AUTH_PUBLIC ) {
+        if( !Auth::check() && is_numeric( config( 'grapher.access.latency' ) ) && (int)config( 'grapher.access.latency' ) === User::AUTH_PUBLIC ) {
             return true;
         }
 
@@ -264,7 +264,7 @@ class Latency extends Graph
         $us = Auth::getUser();
 
         // NB: see above authorisedForAllCustomers()
-        if( is_numeric( config( 'grapher.access.latency' ) ) && config( 'grapher.access.latency' ) === User::AUTH_PUBLIC ) {
+        if( is_numeric( config( 'grapher.access.latency' ) ) && (int)config( 'grapher.access.latency' ) === User::AUTH_PUBLIC ) {
             return $this->allow();
         }
